Basic loading/saving of Inis works correctly (including quotes, comments). Fixed API calls on Windows.

+++ ENIGMAsystem/SHELL/Platforms/Win32/WINDOWSfilemanip.cpp
@@ -59,7 +59,7 @@ std::string ini_read_string(std::string section, std::string key, std::string de
 	return buffer;
 }
 
-int ini_read_real(std::string section, std::string key, int defaultValue)
+float ini_read_real(std::string section, std::string key, float defaultValue)
 {
 	return GetPrivateProfileInt(section.c_str(), key.c_str(), defaultValue, iniFilename.c_str());
 }
@@ -69,7 +69,7 @@ void ini_write_string(std::string section, std::string key, std::string value)
 	WritePrivateProfileString(section.c_str(), key.c_str(), value.c_str(), iniFilename.c_str());
 }
 
-void ini_write_real(std::string section, std::string key, int value)
+void ini_write_real(std::string section, std::string key, float value)
 {
 	std::stringstream ss;
 	ss << value;
